Issue #2920: Move WEBSOCKET_PORT and getWebSocketURL should be abstract
