+++ dev/com.ibm.ws.org.apache.myfaces.2.3/src/org/apache/myfaces/application/ViewHandlerImpl.java
@@ -580,5 +580,12 @@ public class ViewHandlerImpl extends ViewHandler
         return concatenatedStream == null ? Stream.empty() : concatenatedStream;
     }
     
+    @Override
+    public String getWebsocketURL(FacesContext context, String channelAndToken)
+    {
+        String url = context.getExternalContext().getRequestContextPath() + 
+                "/javax.faces.push/"+channelAndToken;
+        return url;
+    }
     
 }
